update architecture search workflow
