+++ docs/src/workflow examples/Basic Neural Architecture Search/Project.toml
@@ -3,4 +3,5 @@ DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 Flux = "587475ba-b771-5e3f-ad9e-33799f191a9c"
 MLJ = "add582a8-e3ab-11e8-2d5e-e98b27df1bc7"
 MLJFlux = "094fc8d1-fd35-5302-93ea-dabda2abf845"
+Optimisers = "3bd65402-5787-11e9-1adc-39752487f4e2"
 RDatasets = "ce6b1742-4840-55fa-b093-852dadbb1d8b"
